backend/requirements.txt
# Core
numpy>=1.23
python-dotenv>=1.0

# Computer Vision
opencv-python>=4.8
Pillow>=10.0

# Deep Learning
tensorflow>=2.13,<2.16
ultralytics>=8.0.0

# API
fastapi>=0.110
uvicorn>=0.27

# Data validation
pydantic>=2.0

# HTTP requests
requests>=2.31

# LLM Clients
groq>=0.5.0

# Background removal (optional but used)
rembg>=2.0.50
onnxruntime>=1.16

# Utilities
python-multipart>=0.0.6